% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gene_variances.R
\name{calculate_gene_variances}
\alias{calculate_gene_variances}
\title{Calculate Gene Variances}
\usage{
calculate_gene_variances(
  gene_names,
  df,
  covariates = c("species", "celltype"),
  verbose = FALSE
)
}
\arguments{
\item{gene_names}{A character vector specifying the names of genes for which variances need to be calculated.}

\item{df}{A data frame containing the relevant data, including expression values of genes and specified covariates.}

\item{covariates}{A character vector specifying the covariates (default: c("species", "celltype")) to include in the linear model.}

\item{verbose}{A logical value indicating whether to display a progress bar (default: FALSE).}
}
\value{
A data frame containing gene names and corresponding variances for residuals, covariate 1, and covariate 2.
}
\description{
This function calculates the variances associated with different sources (residuals and covariates) for a given set of genes in single-nucleus RNA-seq data.
}
\details{
The function fits linear models for each gene using specified covariates and extracts the relevant ANOVA information to calculate variances.
}
\examples{
gene_names <- c("Gene1", "Gene2", "Gene3")
data_frame <- read.csv("snRNAseq_data.csv")
result <- calculate_gene_variances(gene_names, data_frame)

}
